主页面完成,分类页面优化完成,详情页基本完成(待优化),准备开始购物车页面

+++ app/src/main/java/com/lx/android_commerce/mvp/contract/GoodDetailContract.java
@@ -22,7 +22,7 @@ interface GoodDetailContract {
     }
 
     interface Model extends IModel {
-        void requestDetailData(String threePosition, int goods_id, Observer<GoodDetailEntity> observer);
+        void requestDetailData(String threePosition, long goods_id, Observer<GoodDetailEntity> observer);
     }
 
 }
